Elide bad init.go from Go codegen
We correctly skip provider.go, since there are no resources in the
root of the module, but were still emitting init.go. This led to
issue #44, because the init.go code assumes the Provider type exists.
The fix is to also skip emitting init.go.

I would like to add a regression test, but need help figuring out
where that would go, as I didn't see a natural place for it.