Fix Darwin builds (#88)
Due to new upstream dependencies, Darwin builds now require
CGO_ENABLED=1. We still want CGO_ENABLED=0 on other platforms so as to
produce binaries without a libc dependency (especially useful on
Alpine).

Unfortunately the easiest way to make a CGO build work on Darwin is to
build on an actual Mac. For the time being we can fix publishing by just
using a macOS runner with Actions.