Pin the k8s version in the generated resources for node (#121)
Fixes #28 

crd2pulumi uses codegen in a kind of funny way to generate resources
that act as resources of the kubernetes resource plugin. This means the
resource version needs to match a kubernetes plugin version. Generated
resources however, default to the version from the package.json, so if
the user attempts to give the generated package a version, the engine
starts complaining that it can't find a kubernetes plugin with that
version.

This change overrides the generated `getVersion()` helper to point to a
recent kubernetes release so that the engine can find the plugin even if
the user adds a version to the package.json