Resolve package collisions (#153)
## Proposed Changes

This PR ensures the correct CRD versions are used when parsing.
Previously, the Swagger version generated by `builder.BuildOpenAPIV2`
defaulted to `0.1.0`, but we should be pulling the version directly from
the CRD spec itself. This change ensures that SDKs are generated for all
CRD versions. Additionally, this PR incorporates the latest updates from
p-k to prevent package collisions.

## Related Issues (optional)
Closes: #152