Configurable package namespaces/prefixes (#247)
Fixes: #246 

Implemented according to the proposal I posted in #246.

Tested manually with the CertManager CRDs, with and without the new CLI
option, for `nodejs`, `python`, `dotnet` and `java`. As it needs to be,
without the namespace/prefix option, I could verify the default value
was still used. With the new namespace/prefix option, the value was
integrated correctly in the generated files.